requirements.txt
openai==1.14.1
langchain==0.1.12
langchain-openai==0.0.8
google-api-core==2.17.1
google-api-python-client==2.122.0
google-auth==2.28.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.0
langchain-anthropic==0.1.4
python-dotenv==1.0.1